fix(runtime): make autonomous adoption replayable
